+++ Resources/assets/js/journalEntryPreventDoubleSubmission.js
@@ -1,0 +1,11 @@
+$(() => {
+    let journalEntryForm = $('#journal-entry-form');
+
+    journalEntryForm.on('submit', function (e) {
+        if (journalEntryForm.data('submitted')) {
+            e.preventDefault();
+        } else {
+            journalEntryForm.data('submitted', true);
+        }
+    });
+});